Removed all the code generation part
For the compiler lab assignment, only the parser is needed, so removed 
all the code generation I was doing. Now I can more easily finish the 
bison spec and parse, without worrying about all the complexity related 
to code generation. Of course, all this code is still available in 
master which I will use when I want to develop this compiler further.