Changed the global variable block_stats to a local one
If there are nested blocks, this variable will get used by the other 
block. It will then free() it and set it to NULL. So, the outer block 
gets its data corrupted and leaked.

Fix this by creating a list of nodes for each block. The blockdast: stat 
rule creates it and the blockdash: blockdash stat rule then adds in that 
list. This way, nested block stats can be used.